RT #73556 (failing) test case
RT #73556: Test::MockObject::Extends->mock should duplicate prototype
This test case shows the problem: a 'Prototype mismatch' warning is
raised when a method declared with a prototype is replaced with a mock.
The problem comes from the mock() implementation that wraps the mock in
a sub without prototype.
The same kind of problem occurs for subs with attributes (lvalue for
example).
This work has been sponsored by Ijenko.